Standardize html cleanup
