+++ learning_resources/etl/mitxonline.py
@@ -19,12 +19,12 @@ from learning_resources.constants import (
 )
 from learning_resources.etl.constants import ETLSource
 from learning_resources.etl.utils import (
-    clean_data,
     extract_valid_department_from_id,
     generate_course_numbers_json,
     parse_certification,
     transform_topics,
 )
+from main.utils import clean_data
 
 log = logging.getLogger(__name__)
