+++ learning_resources/etl/podcast.py
@@ -12,9 +12,8 @@ from requests.exceptions import HTTPError
 
 from learning_resources.constants import LearningResourceType
 from learning_resources.etl.constants import ETLSource
-from learning_resources.etl.utils import clean_data
 from learning_resources.models import PodcastEpisode
-from main.utils import frontend_absolute_url, now_in_utc
+from main.utils import clean_data, frontend_absolute_url, now_in_utc
 
 CONFIG_FILE_REPO = "mitodl/open-podcast-data"
 CONFIG_FILE_FOLDER = "podcasts"
