+++ learning_resources/etl/youtube.py
@@ -22,9 +22,8 @@ from learning_resources.constants import LearningResourceType, OfferedBy, Platfo
 from learning_resources.etl.constants import ETLSource
 from learning_resources.etl.exceptions import ExtractException
 from learning_resources.etl.loaders import update_index
-from learning_resources.etl.utils import clean_data
 from learning_resources.models import LearningResource
-from main.utils import now_in_utc
+from main.utils import clean_data, now_in_utc
 
 CONFIG_FILE_REPO = "mitodl/open-video-data"
 CONFIG_FILE_FOLDER = "youtube"
