+++ news_events/etl/medium_mit_news_test.py
@@ -71,8 +71,8 @@ def test_transform_items(mocker, medium_mit_rss_data):
         "guid": "https://medium.com/p/a685e2b89c6a",
         "title": "Meet 8 MIT women faculty who teach MITx courses and lead cutting-edge research",
         "url": "https://medium.com/open-learning/meet-8-mit-women-faculty",
-        "summary": "<h4>Celebrating Women\u2019s History Month with MIT women\u2019s truncated</h4>",
-        "content": "<h4>Celebrating Women\u2019s History Month with MIT women\u2019s truncated</h4>",
+        "summary": "Celebrating Women\u2019s History Month with MIT women\u2019s truncated",
+        "content": "Celebrating Women\u2019s History Month with MIT women\u2019s truncated",
         "image": None,
         "detail": {
             "authors": ["MIT Open Learning"],
