+++ news_events/etl/sloan_exec_news_test.py
@@ -43,7 +43,9 @@ def test_transform():
     items = list(source["items"])
     assert len(items) == 20
     assert items[0]["detail"]["publish_date"] == "2024-03-25T16:29:25.000Z"
-    assert items[0]["title"] == "Cybersecurity Resiliency is More Than Protection"
+    assert (
+        items[0]["title"] == "Cybersecurity Resiliency & Stuff is More Than Protection"
+    )
     assert items[1]["title"] == (
         "From blockchain to virtual reality and beyond: Joshua Gayman\u2019s experience and Executive Certificate journey at MIT"
     )
@@ -56,6 +58,12 @@ def test_transform():
         "Trending Blog Posts",
     ]
     assert items[3]["detail"]["authors"] == ['MIT Professor Fiona "OG" Murray']
+    assert items[3]["summary"].startswith(
+        "<p>Despite the many advances women have made in the workforce over"
+    )
+    assert items[3]["content"].startswith(
+        "<p>During Women\u2019s History Month, it is important"
+    )
 
 
 def test_transform_no_posts(mocker):
