+++ test_html/test_medium_mit_news.rss
@@ -26,7 +26,7 @@
             <dc:creator><![CDATA[MIT Open Learning]]></dc:creator>
             <pubDate>Fri, 15 Mar 2024 13:42:36 GMT</pubDate>
             <atom:updated>2024-03-15T13:58:19.238Z</atom:updated>
-            <content:encoded><![CDATA[<h4>Celebrating Women’s History Month with MIT women’s truncated</h4>]]></content:encoded>
+            <content:encoded><![CDATA[<h4>Celebrating Women’s History Month with <a href=""https://www.mit.edu>MIT</a> women’s truncated</h4>]]></content:encoded>
         </item>
         <item>
             <title><![CDATA[Ten books from MIT faculty to expand your knowledge of teaching, learning, and technology]]></title>
