+++ test_json/ol_events_output.json
@@ -7,8 +7,8 @@
     "alt": "students raising their hands",
     "description": "Students Raising Their Hands"
   },
-  "summary": "The MicroMasters® program in Data, Economics, and Design of Policy (DEDP) equips learners with the practical skills and theoretical knowledge to tackle some of the most pressing challenges around the world.&nbsp;Through eight courses taught by MIT faculty, learners will gain a strong foundation in microeconomics, development economics, probability, and statistics, and engage with cutting-edge research in the field. The program is unique in its focus on running randomized evaluations to assess the effectiveness of social programs and its emphasis on hands-on skills in data analysis.&nbsp;All course content is available online for free. Learners who wish to earn certificates in the courses can pay a fee to take proctored exams. Learners who complete three core and two elective courses will earn a DEDP MicroMasters program credential in their respective track and become eligible to apply to the on-campus master’s in DEDP at MIT and other graduate programs at participating global pathway schools.&nbsp;Attend our 60-minute webinar to learn more and to have your questions answered live by the DEDP program team!&nbsp;Can't attend the live webinar? Register anyway! A recording will be sent to all registrants.Register &gt;&gt;",
-  "content": "The MicroMasters® program in Data, Economics, and Design of Policy (DEDP) equips learners with the practical skills and theoretical knowledge to tackle some of the most pressing challenges around the world.&nbsp;Through eight courses taught by MIT faculty, learners will gain a strong foundation in microeconomics, development economics, probability, and statistics, and engage with cutting-edge research in the field. The program is unique in its focus on running randomized evaluations to assess the effectiveness of social programs and its emphasis on hands-on skills in data analysis.&nbsp;All course content is available online for free. Learners who wish to earn certificates in the courses can pay a fee to take proctored exams. Learners who complete three core and two elective courses will earn a DEDP MicroMasters program credential in their respective track and become eligible to apply to the on-campus master’s in DEDP at MIT and other graduate programs at participating global pathway schools.&nbsp;Attend our 60-minute webinar to learn more and to have your questions answered live by the DEDP program team!&nbsp;Can't attend the live webinar? Register anyway! A recording will be sent to all registrants.Register &gt;&gt;",
+  "summary": "<p>The MicroMasters® program in Data, Economics, and Design of Policy (DEDP) equips learners with the practical skills and theoretical knowledge to tackle some of the most pressing challenges around the world.&nbsp;</p><p>Through eight courses taught by MIT faculty, learners will gain a strong foundation in microeconomics, development economics, probability, and statistics, and engage with cutting-edge research in the field. The program is unique in its focus on running randomized evaluations to assess the effectiveness of social programs and its emphasis on hands-on skills in data analysis.&nbsp;</p><p>All course content is available online for free. Learners who wish to earn certificates in the courses can pay a fee to take proctored exams. Learners who complete three core and two elective courses will earn a DEDP MicroMasters program credential in their respective track and become eligible to apply to the on-campus master’s in DEDP at MIT and other graduate programs at participating global pathway schools.&nbsp;</p><p>Attend our 60-minute webinar to learn more and to have your questions answered live by the DEDP program team!&nbsp;</p><p>Can't attend the live webinar? Register anyway! A recording will be sent to all registrants.</p><p><strong>Register &gt;&gt;</strong></p>",
+  "content": "<p>The MicroMasters® program in Data, Economics, and Design of Policy (DEDP) equips learners with the practical skills and theoretical knowledge to tackle some of the most pressing challenges around the world.&nbsp;</p><p>Through eight courses taught by MIT faculty, learners will gain a strong foundation in microeconomics, development economics, probability, and statistics, and engage with cutting-edge research in the field. The program is unique in its focus on running randomized evaluations to assess the effectiveness of social programs and its emphasis on hands-on skills in data analysis.&nbsp;</p><p>All course content is available online for free. Learners who wish to earn certificates in the courses can pay a fee to take proctored exams. Learners who complete three core and two elective courses will earn a DEDP MicroMasters program credential in their respective track and become eligible to apply to the on-campus master’s in DEDP at MIT and other graduate programs at participating global pathway schools.&nbsp;</p><p>Attend our 60-minute webinar to learn more and to have your questions answered live by the DEDP program team!&nbsp;</p><p>Can't attend the live webinar? Register anyway! A recording will be sent to all registrants.</p><p><strong>Register &gt;&gt;</strong></p>",
   "detail": {
     "location": ["Online"],
     "audience": ["Faculty", "MIT Community", "Public", "Students"],
